updated file

+++ MA4git/MA4_files/person.cpp
@@ -24,7 +24,7 @@ void Person::setAge(int a){
 	}
 
 double Person::getDecades(){
-	return age/10;
+	return double(age)/10;
 	}
